Added new funcionality - for Boss class
New interface created - BossShotEntity - it is a class for fireballs
which boss can shoot out against the player. All shots are stored in
Controller class in ArrayList. The Controller can add and remove
fireballs. New method for detect collisions with this BossShotEntity
class is added in Physics class. New .wav files added in resources - now
when boss shoots or boss's shot hits a target, a sound is produced.
Minimum distance for creating a shot and  limitations for how far is the
previous shot from current boss position, are set.
Unused fields are removed from Game class.