Health restorer Class is now present
New HealthRestorer class is added in the application. The player can
interact with this class and restore the provided health quantity.
The class is created in the controller in a new collection named
"healthRestorersEntities". When an enemy which is hunter is killed, in
its placed is left a new Health restorer.  The life if this class is
within 250 ticks of the game.